refactor: semantic separation of sduwh with sdu
