+++ lib/v2/sdu/extractor/index.js
@@ -1,6 +1,6 @@
 module.exports = async (link, ctx) => {
     if (link.startsWith('https://xinwen.wh.sdu.edu.cn/')) {
-        return await require('./news')(link, ctx);
+        return await require('./wh/news')(link, ctx);
     }
     if (link.startsWith('https://www.view.sdu.edu.cn/')) {
         return await require('./view')(link, ctx);
