+++ lib/v2/sdu/extractor/wh/news.js
