+++ lib/v2/sdu/maintainer.js
@@ -1,3 +1,3 @@
 module.exports = {
-    '/news/:column?': ['kxxt'],
+    '/wh/news/:column?': ['kxxt'],
 };
