+++ lib/v2/sdu/radar.js
@@ -1,13 +1,13 @@
-const { news } = require('./data');
+const { wh } = require('./data');
 
 module.exports = {
     'sdu.edu.cn': {
         _name: '山东大学',
-        'xinwen.wh': Object.entries(news.columns).map(([, value]) => ({
+        'xinwen.wh': Object.entries(wh.news.columns).map(([, value]) => ({
             title: value.name,
-            docs: news.docs,
-            source: news.source,
-            target: '/sdu/wh/' + news.getTarget(value.url),
+            docs: wh.news.docs,
+            source: wh.news.source,
+            target: '/sdu/wh' + wh.news.getTarget(value.url),
         })),
     },
 };
