+++ lib/v2/sdu/router.js
@@ -1,3 +1,3 @@
 module.exports = function (router) {
-    router.get('/wh/news/:column?', require('./news'));
+    router.get('/wh/news/:column?', require('./wh/news'));
 };
