+++ lib/v2/sdu/wh/news.js
@@ -1,7 +1,7 @@
 const got = require('@/utils/got');
 const cheerio = require('cheerio');
-const data = require('./data').news;
-const extractor = require('./extractor');
+const data = require('../data').wh.news;
+const extractor = require('../extractor');
 const { parseDate } = require('@/utils/parse-date');
 
 module.exports = async (ctx) => {
@@ -32,7 +32,7 @@ module.exports = async (ctx) => {
 
     ctx.state.data = {
         title: `${data.name} ${data.columns[column].name}`,
-        link: 'https://xinwen.wh.sdu.edu.cn/',
+        link: data.url,
         item: out,
     };
 };
